Move SQLite store implementations to own module
