+++ src/stores/mod.rs
@@ -1,12 +1,13 @@
 //! Contains traits and implementations for objects that store the domain [models](crate::models).
 
-pub mod balance;
-pub mod category;
-pub mod sql_store;
-pub mod transaction;
-pub mod user;
+mod balance;
+mod category;
+mod transaction;
+mod user;
+
+pub mod sqlite;
 
 pub use balance::BalanceStore;
-pub use category::{CategoryStore, SQLiteCategoryStore};
-pub use transaction::{SQLiteTransactionStore, TransactionStore};
-pub use user::{SQLiteUserStore, UserStore};
+pub use category::CategoryStore;
+pub use transaction::{TransactionStore, TransactionQuery, SortOrder};
+pub use user::UserStore;
